+++ src/integrator.c
@@ -267,7 +267,7 @@ void reb_simulation_update_acceleration(struct reb_simulation* r){
         if (r->integrator==REB_INTEGRATOR_MERCURIUS){
             struct reb_particle* restrict const particles = r->particles;
             struct reb_particle* restrict const backup = r->ri_mercurius.particles_backup_additional_forces;
-            for (unsigned int i=0;i<r->N;i++){
+            for (size_t i=0;i<r->N;i++){
                 particles[i].x = backup[i].x;
                 particles[i].y = backup[i].y;
                 particles[i].z = backup[i].z;
@@ -279,7 +279,7 @@ void reb_simulation_update_acceleration(struct reb_simulation* r){
         if (r->integrator==REB_INTEGRATOR_TRACE){
             struct reb_particle* restrict const particles = r->particles;
             struct reb_particle* restrict const backup = r->ri_trace.particles_backup_additional_forces;
-            for (unsigned int i=0;i<r->N;i++){
+            for (size_t i=0;i<r->N;i++){
                 particles[i].x = backup[i].x;
                 particles[i].y = backup[i].y;
                 particles[i].z = backup[i].z;
